Add policy files for topic and bookmark, add user_id to bookmark

+++ app/controllers/bookmarks_controller.rb
@@ -2,7 +2,7 @@ class BookmarksController < ApplicationController
 
   def destroy
     @bookmark = Bookmark.find(params[:id])
-    
+    authorize @bookmark
     if @bookmark.destroy
       flash[:notice] = "Bookmark was deleted"
     else
